Add data folder and storage.txt and add reading storage.txt on startup
