+++ data/storage.txt
@@ -1,0 +1,3 @@
+T|A
+D|B|Today
+E|C|Noon|Night
