+++ src/main/java/dodo/Event.java
@@ -1,3 +1,5 @@
+package dodo;
+
 public class Event extends Task{
     protected String start;
     protected String end;
