+++ src/main/java/dodo/Task.java
@@ -1,3 +1,5 @@
+package dodo;
+
 public class Task {
     protected String description;
     protected boolean isDone;
