+++ src/main/java/dodo/Todo.java
@@ -1,3 +1,5 @@
+package dodo;
+
 public class Todo extends Task{
     public Todo(String description) {
         super(description);
