+++ text-ui-test/EXPECTED.TXT
@@ -30,18 +30,18 @@ You now have 2 task(s).
 Added this to your list:
 [E][ ] c (from: 1 to: 2)
 You now have 3 task(s).
-Deadline commands needs to be structured as follows:
+dodo.Deadline commands needs to be structured as follows:
 deadline 'name' /by 'time'
-Event commands needs to be structured as follows:
+dodo.Event commands needs to be structured as follows:
 event 'name' /from 'start time' /to 'end time'
 Here are your tasks:
 1. [T][ ] a
 2. [D][ ] b (by: 1)
 3. [E][ ] c (from: 1 to: 2)
-Task number 0 doesn't exist dodohead!
-Task number 4 doesn't exist dodohead!
-Task number 0 doesn't exist dodohead!
-Task number 4 doesn't exist dodohead!
+dodo.Task number 0 doesn't exist dodohead!
+dodo.Task number 4 doesn't exist dodohead!
+dodo.Task number 0 doesn't exist dodohead!
+dodo.Task number 4 doesn't exist dodohead!
 Marked as done:
 [T][X] a
 Following task is already marked done:
